add --invert option to invert pulse (0=HIGH, 1=LOW)
New feature: --invert

This is useful for common anode LEDs where the pin should sink current to
light the LED.  Instead of the output being held LOW and going HIGH for the
pulse duration, the output is held HIGH and is driven LOW for the pulse
duration.  This the pins are left HIGH when exiting as well.